CORE-5965: Package restructuring on server-side.
