+++ applications/belphegor/src/main/resources/belphegor-base.xml
@@ -56,7 +56,7 @@
 
     <bean id="singleLogoutFilter" class="org.jasig.cas.client.session.SingleSignOutFilter" />
 
-    <bean id="logoutHandler" class="org.iplantc.de.server.CasLogoutSuccessHandler">
+    <bean id="logoutHandler" class="org.iplantc.de.server.auth.CasLogoutSuccessHandler">
         <property name="logoutUrl" value="${org.iplantc.belphegor.cas.base-url}/logout" />
         <property name="defaultRedirectUrl" value="${org.iplantc.belphegor.cas.server-name}/belphegor/logged-out" />
         <property name="redirectUrlSelectorName" value="reason" />
@@ -94,7 +94,7 @@
         <task:scheduled ref="proxyGrantingTicketStorage" method="cleanUp" fixed-delay="300000"/>
     </task:scheduled-tasks>
 
-    <bean id="authenticationUserDetailsService" class="org.iplantc.de.server.CasGroupUserDetailsService">
+    <bean id="authenticationUserDetailsService" class="org.iplantc.de.server.auth.CasGroupUserDetailsService">
         <constructor-arg index="0" value="${org.iplantc.belphegor.cas.group-attribute-name}" />
     </bean>
 </beans>
