+++ applications/de-webapp/src/main/java/org/iplantc/de/server/oauth/CasOAuthCallbackServlet.java
@@ -1,6 +1,6 @@
 package org.iplantc.de.server.oauth;
 
-import org.iplantc.de.server.CasUrlConnector;
+import org.iplantc.de.server.auth.CasUrlConnector;
 import org.iplantc.de.server.DiscoveryEnvironmentProperties;
 
 public class CasOAuthCallbackServlet  extends OAuthCallbackServlet {
