+++ applications/de-webapp/src/main/webapp/WEB-INF/web.xml
@@ -14,7 +14,7 @@
     <!-- Servlets -->
     <servlet>
         <servlet-name>sessionInitializationServlet</servlet-name>
-        <servlet-class>org.iplantc.de.server.CasSessionInitializationServlet</servlet-class>
+        <servlet-class>org.iplantc.de.server.auth.CasSessionInitializationServlet</servlet-class>
         <init-param>
             <param-name>initialPage</param-name>
             <param-value>Discoveryenvironment.jsp#workspace</param-value>
@@ -22,7 +22,7 @@
     </servlet>
     <servlet>
         <servlet-name>logoutSuccessServlet</servlet-name>
-        <servlet-class>org.iplantc.de.server.CasLogoutServlet</servlet-class>
+        <servlet-class>org.iplantc.de.server.auth.CasLogoutServlet</servlet-class>
         <init-param>
             <param-name>propertyNamePrefix</param-name>
             <param-value>org.iplantc.discoveryenvironment</param-value>
@@ -30,7 +30,7 @@
     </servlet>
     <servlet>
         <servlet-name>loggedOutServlet</servlet-name>
-        <servlet-class>org.iplantc.de.server.CasLoggedOutServlet</servlet-class>
+        <servlet-class>org.iplantc.de.server.auth.CasLoggedOutServlet</servlet-class>
         <init-param>
             <param-name>propertyNamePrefix</param-name>
             <param-value>org.iplantc.discoveryenvironment</param-value>
@@ -46,7 +46,7 @@
     </servlet>
     <servlet>
         <servlet-name>sessionManagementServlet</servlet-name>
-        <servlet-class>org.iplantc.de.server.SessionManagementServiceImpl</servlet-class>
+        <servlet-class>org.iplantc.de.server.service.SessionManagementServiceImpl</servlet-class>
     </servlet>
     <servlet>
         <servlet-name>propertyServlet</servlet-name>
