+++ modules/de-common-module/src/main/java/org/iplantc/de/server/auth/DESecurityConstants.java
@@ -1,4 +1,4 @@
-package org.iplantc.de.server;
+package org.iplantc.de.server.auth;
 
 import java.util.Collection;
 import java.util.HashMap;
