+++ modules/de-common-module/src/main/java/org/iplantc/de/server/auth/UrlConnector.java
@@ -1,4 +1,4 @@
-package org.iplantc.de.server;
+package org.iplantc.de.server.auth;
 
 import org.apache.http.client.methods.*;
