+++ modules/de-common-module/src/main/java/org/iplantc/de/server/service/DEServiceImpl.java
@@ -1,15 +1,15 @@
 package org.iplantc.de.server.service;
 
-import org.iplantc.de.server.CasUrlConnector;
-import org.iplantc.de.server.DESecurityConstants;
+import org.iplantc.de.server.auth.CasUrlConnector;
+import org.iplantc.de.server.auth.DESecurityConstants;
 import org.iplantc.de.server.DEServiceInputStream;
 import org.iplantc.de.server.MultipartBodyFactory;
 import org.iplantc.de.server.ServiceCallResolver;
-import org.iplantc.de.server.UrlConnector;
-import org.iplantc.de.shared.AuthenticationException;
+import org.iplantc.de.server.auth.UrlConnector;
+import org.iplantc.de.shared.exceptions.AuthenticationException;
 import org.iplantc.de.shared.services.DEService;
-import org.iplantc.de.shared.HttpException;
-import org.iplantc.de.shared.HttpRedirectException;
+import org.iplantc.de.shared.exceptions.HttpException;
+import org.iplantc.de.shared.exceptions.HttpRedirectException;
 import org.iplantc.de.shared.services.BaseServiceCallWrapper;
 import org.iplantc.de.shared.services.HTTPPart;
 import org.iplantc.de.shared.services.MultiPartServiceWrapper;
