+++ modules/de-common-module/src/main/java/org/iplantc/de/shared/exceptions/ConfluenceException.java
@@ -1,4 +1,4 @@
-package org.iplantc.de.shared;
+package org.iplantc.de.shared.exceptions;
 
 import org.iplantc.de.shared.services.ConfluenceService;
