+++ modules/de-common-module/src/main/java/org/iplantc/de/shared/exceptions/HttpRedirectException.java
@@ -1,6 +1,4 @@
-package org.iplantc.de.shared;
-
-import java.io.IOException;
+package org.iplantc.de.shared.exceptions;
 
 /**
  * Thrown by the service dispatcher when an HTTP redirect is received.
